Fix checking for harvest objects in fetch queue.
The fetch queue is stored in redis as a list of harvest object ids. The previous code checked for individual harvest object ids in redis, and always found none. Therefore, every time the harvester run command was run, all ids of harvest objects in the WAITING state in the DB were re-added to the fetch queue.